fixed major bugs

+++ 11_Final_EDtech_project/backend/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/Subsection")
 // CREATE a new section
 exports.createSection = async (req, res) => {
 	try {
@@ -54,15 +55,27 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId , courseId } = req.body;
 		const section = await Section.findByIdAndUpdate(
 			sectionId,
 			{ sectionName },
 			{ new: true }
 		);
+
+		const course = await Course.findById(courseId)
+		.populate( {
+			path:"courseContent",
+			populate: {
+				path: "subSection",
+			}
+		})
+		.exec()
+
+
 		res.status(200).json({
 			success: true,
 			message: section,
+			data:course,
 		});
 	} catch (error) {
 		console.error("Error updating section:", error);
@@ -73,22 +86,82 @@ exports.updateSection = async (req, res) => {
 	}
 };
 
-// DELETE a section
+// // DELETE a section
+// exports.deleteSection = async (req, res) => {
+// 	try {
+// 		const { sectionId } = req.body;
+// 		await Section.findByIdAndDelete(sectionId);
+
+// 		//update the course also by removing the sectionID from its Sections 
+// 		res.status(200).json({
+// 			success: true,
+// 			message: "Section deleted",
+// 		});
+// 	} catch (error) {
+// 		console.error("Error deleting section:", error);
+// 		res.status(500).json({
+// 			success: false,
+// 			message: "Internal server error",
+// 		});
+// 	}
+// };
+
 exports.deleteSection = async (req, res) => {
 	try {
-		const { sectionId } = req.body;
+
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
 		await Section.findByIdAndDelete(sectionId);
 
-		//update the course also by removing the sectionID from its Sections 
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
 		res.status(200).json({
-			success: true,
-			message: "Section deleted",
+			success:true,
+			message:"Section deleted",
+			data:course
 		});
+
+
+
+
+
+		// //HW -> req.params -> test
+		// const { sectionId } = req.params;
+		// await Section.findByIdAndDelete(sectionId);
+		// //HW -> Course ko bhi update karo
+		// res.status(200).json({
+		// 	success: true,
+		// 	message: "Section deleted",
+		// });
 	} catch (error) {
 		console.error("Error deleting section:", error);
-		res.status(500).json({
+		res.status(403).json({
+			// error:error.message,
 			success: false,
-			message: "Internal server error",
+			message: error.message,
 		});
 	}
 };
